router: fold `local` into the model it is about
`router local` was named after the plane it reached rather than the thing
it reached about. A user with a question about a model had to know the
answer lived behind an application, and know that application's Olares
app id to ask it — while `router model` and `router list` next door took
the model's own name.

The eight verbs become verbs on `model`: status, progress, retry and
restart directly, and gpu, perf, config and endpoints under `model diag`,
which keeps the four questions about a model that already works from
crowding the ones about whether it works at all. `local spec set` and
`local spec file` join `model spec`, where `edit` and `set` now sit side
by side and the help says plainly that one merges through Router and the
other replaces the document at the application.

The app id survives as `--app`, and is the reason this is a merge rather
than a rename. Naming a model costs a lookup in Router, and a model that
is not answering is exactly when Router may be unreachable or wrong about
which application serves it; `--app` skips the lookup entirely, so the
diagnostic stays available in the case it exists for. `model spec show`
and `model restart` therefore carry both roads, choosing on the flag.

`router local` stays as a hidden deprecated tree. Its verbs are the same
constructors built with the old positional shape, so there is one
implementation rather than two that drift.
