+++ cli/cmd/ctl/router/model_diag.go
@@ -7,22 +7,63 @@ import (
 	"io"
 	"net/url"
 	"os"
+	"sort"
+	"strings"
 
 	"github.com/spf13/cobra"
 
-	"github.com/beclab/Olares/cli/pkg/cliutil"
 	"github.com/beclab/Olares/cli/pkg/cmdutil"
 	"github.com/beclab/Olares/cli/pkg/utils"
 )
 
-// `router local gpu`  — GET  /api/diag/gpu
-// `router local perf` — POST /api/diag/perf, GET /api/diag/perf/last
+// `router model diag …` — the four questions about a local model that are not
+// about whether it works.
 //
-// Both answer "the model replies, but slowly", which Router cannot see: from
-// the gateway a model on the CPU and a model on the GPU differ only in latency.
-// The figures come from the engine's own introspection rather than from
+// gpu       GET  /api/diag/gpu
+// perf      POST /api/diag/perf, GET /api/diag/perf/last
+// config    GET  /api/config
+// endpoints GET  /api/endpoints
+//
+// gpu and perf both answer "the model replies, but slowly", which Router cannot
+// see: from the gateway a model on the CPU and a model on the GPU differ only in
+// latency. The figures come from the engine's own introspection rather than from
 // nvidia-smi, so what is reported varies by engine and an absent field means
 // "this engine does not say" rather than zero.
+//
+// config and endpoints answer the question underneath that one — what this
+// deployment was actually asked to be, and which routes it therefore serves.
+//
+// These are separated from the lifecycle verbs because they answer a different
+// question. `model status` says whether a model is usable; these say why a model
+// that is usable behaves the way it does.
+
+// newModelDiagCommand groups them under one noun rather than four more verbs on
+// `model`, which already carries the ones a person reaches for daily.
+func newModelDiagCommand(f *cmdutil.Factory, how localAddressing) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "diag",
+		Short: "why a working local model behaves the way it does",
+		Long: `Look inside a model application that is running.
+
+  diag gpu <model>        how much of the model is resident on the GPU
+  diag perf <model>       time to first token and throughput, measured
+  diag config <model>     the effective configuration, secrets redacted
+  diag endpoints <model>  which routes this deployment actually serves
+
+Whether a model is usable at all is "router model status". These are for the
+one that is: slow, on the wrong device, or missing a route you expected.
+
+Every verb takes a model, and --app to name the application directly instead of
+asking Router which one serves it.
+`,
+	}
+	cmd.SilenceUsage = true
+	cmd.AddCommand(newModelGPUCommand(f, how))
+	cmd.AddCommand(newModelPerfCommand(f, how))
+	cmd.AddCommand(newModelConfigCommand(f, how))
+	cmd.AddCommand(newModelEndpointsCommand(f, how))
+	return cmd
+}
 
 type gpuResidency struct {
 	VRAMBytes            int64    `json:"vram_bytes,omitempty"`
@@ -75,13 +116,14 @@ func (g gpuReport) placement() string {
 	return "not reported by this engine"
 }
 
-func newLocalGPUCommand(f *cmdutil.Factory) *cobra.Command {
+func newModelGPUCommand(f *cmdutil.Factory, how localAddressing) *cobra.Command {
 	var (
 		output string
 		native bool
 	)
+	target := newModelTarget(how)
 	cmd := &cobra.Command{
-		Use:   "gpu <app>",
+		Use:   "gpu " + target.arg(),
 		Short: "how much of the model is resident on the GPU",
 		Long: `Show where a model's weights actually are.
 
@@ -95,10 +137,9 @@ KV-cache use is how busy the model is right now. An idle model fully on the GPU
 reports 0% cache use.
 
 Examples:
-  olares-cli router local gpu llamacppqwen3627bggufv3
-  olares-cli router local gpu llamacppqwen3627bggufv3 --engine-native -o json
+  olares-cli router model diag gpu qwen3-4b
+  olares-cli router model diag gpu --app llamacppqwen3627bggufv3 --engine-native -o json
 `,
-		Args: cobra.ExactArgs(1),
 		RunE: func(c *cobra.Command, args []string) error {
 			ctx := c.Context()
 			if ctx == nil {
@@ -108,7 +149,7 @@ Examples:
 			if err != nil {
 				return err
 			}
-			li, err := openLocal(ctx, f, args[0])
+			li, err := target.open(ctx, f, args)
 			if err != nil {
 				return err
 			}
@@ -131,6 +172,7 @@ Examples:
 			return renderGPUReport(os.Stdout, li, &rep)
 		},
 	}
+	target.bind(cmd)
 	cmd.Flags().BoolVar(&native, "engine-native", false,
 		"include the engine's raw introspection dump (JSON output only)")
 	addOutputFlag(cmd, &output)
@@ -180,7 +222,7 @@ func renderGPUReport(w io.Writer, li *llmInit, r *gpuReport) error {
 	if r.placement() == "not reported by this engine" {
 		_, err := fmt.Fprintf(w, "\nThis engine reports no residency figures, so where the weights sit "+
 			"cannot be read from here. What it was asked to do is in its launch flags: "+
-			"`olares-cli router local config %s`.\n", li.AppName)
+			"`olares-cli router model diag config --app %s`.\n", li.AppName)
 		return err
 	}
 	return nil
@@ -220,7 +262,7 @@ type perfRunRequest struct {
 	WithoutThink *bool  `json:"without_think,omitempty"`
 }
 
-func newLocalPerfCommand(f *cmdutil.Factory) *cobra.Command {
+func newModelPerfCommand(f *cmdutil.Factory, how localAddressing) *cobra.Command {
 	var (
 		output       string
 		last         bool
@@ -229,8 +271,9 @@ func newLocalPerfCommand(f *cmdutil.Factory) *cobra.Command {
 		withThink    bool
 		onlyThink    bool
 	)
+	target := newModelTarget(how)
 	cmd := &cobra.Command{
-		Use:   "perf <app>",
+		Use:   "perf " + target.arg(),
 		Short: "time to first token and throughput, measured",
 		Long: `Measure a model by using it.
 
@@ -248,11 +291,10 @@ second is the steady state.
 is in memory, so a restarted application has none.
 
 Examples:
-  olares-cli router local perf llamacppqwen3627bggufv3
-  olares-cli router local perf llamacppqwen3627bggufv3 --last
-  olares-cli router local perf llamacppqwen3627bggufv3 --decode-tokens 128 --with-think
+  olares-cli router model diag perf qwen3-4b
+  olares-cli router model diag perf qwen3-4b --last
+  olares-cli router model diag perf --app llamacppqwen3627bggufv3 --decode-tokens 128 --with-think
 `,
-		Args: cobra.ExactArgs(1),
 		RunE: func(c *cobra.Command, args []string) error {
 			ctx := c.Context()
 			if ctx == nil {
@@ -266,7 +308,7 @@ Examples:
 				return fmt.Errorf("--last reads the run that already happened, so the knobs that shape a " +
 					"new one cannot apply to it; drop --last to measure again")
 			}
-			li, err := openLocal(ctx, f, args[0])
+			li, err := target.open(ctx, f, args)
 			if err != nil {
 				return err
 			}
@@ -293,6 +335,7 @@ Examples:
 			return renderPerfReport(os.Stdout, li, &rep, last)
 		},
 	}
+	target.bind(cmd)
 	cmd.Flags().BoolVar(&last, "last", false, "read the previous run instead of starting one")
 	cmd.Flags().IntVar(&decodeTokens, "decode-tokens", 0, "tokens to sample while measuring decode (default 64, max 256)")
 	cmd.Flags().StringVar(&prompt, "prompt", "", "use this prompt for the prefill pass instead of the built-in one")
@@ -355,10 +398,10 @@ func diagErr(err error) error {
 	switch {
 	case re.Status == 404:
 		return fmt.Errorf("%w\nThis application does not serve the diagnostic routes; "+
-			"`olares-cli router local endpoints <app>` lists what it does serve", err)
+			"`olares-cli router model diag endpoints <model>` lists what it does serve", err)
 	case re.Status == 503:
 		return fmt.Errorf("%w\nThe engine did not answer its own introspection probe, which usually means "+
-			"it is not up yet; `olares-cli router local status <app>` says which phase it is in", err)
+			"it is not up yet; `olares-cli router model status <model>` says which phase it is in", err)
 	}
 	return err
 }
@@ -371,7 +414,7 @@ func perfRunErr(err error) error {
 	switch re.Code {
 	case "engine_not_ready":
 		return fmt.Errorf("%w\nMeasuring an engine that is still starting would record timings for work it "+
-			"is not doing yet; `olares-cli router local progress <app> --watch` follows it to ready", err)
+			"is not doing yet; `olares-cli router model progress <model> --watch` follows it to ready", err)
 	case "perf_run_in_progress":
 		return fmt.Errorf("%w\nOne run at a time per application. Wait for the other to finish, or read it "+
 			"with --last once it has", err)
@@ -388,38 +431,25 @@ func perfLastErr(err error) error {
 	return diagErr(err)
 }
 
-func newLocalRetryCommand(f *cmdutil.Factory) *cobra.Command {
-	var (
-		output string
-		force  bool
-		level  string
-	)
+func newModelConfigCommand(f *cmdutil.Factory, how localAddressing) *cobra.Command {
+	var output string
+	target := newModelTarget(how)
 	cmd := &cobra.Command{
-		Use:   "retry <app>",
-		Short: "re-enter the download and load loop now",
-		Long: `Tell a model application to try again immediately.
+		Use:   "config " + target.arg(),
+		Short: "the effective configuration, secrets redacted",
+		Long: `Show the configuration a model application is actually running with.
 
-An application that failed, or is waiting between attempts, sleeps before its
-next one. This wakes it. It is the right response to a download that failed for a
-reason since fixed — a network outage, a token that has been replaced, a disk
-that had no room.
+Secrets are masked by the Model Console before they leave it — a token becomes a
+length, a source URL keeps only its scheme and host — so this is safe to paste
+into a bug report.
 
---force re-downloads everything, ignoring the manifest's account of what is
-already good. That is a full re-pull, potentially many gigabytes, and it is the
-only thing that gets past a manifest that wrongly believes the files are intact.
-
---level chooses how thoroughly this one attempt verifies: size is quick, sha256
-reads every byte.
-
-Calls are rate limited by the application, so a loop of these earns a refusal
-rather than faster progress.
+It is the answer to "why is it downloading from there" and "which engine flags
+are in force", both of which are set at install time and easy to misremember.
 
 Examples:
-  olares-cli router local retry llamacppqwen3627bggufv3
-  olares-cli router local retry llamacppqwen3627bggufv3 --level sha256
-  olares-cli router local retry llamacppqwen3627bggufv3 --force
+  olares-cli router model diag config qwen3-4b
+  olares-cli router model diag config --app llamacppqwen3627bggufv3 -o json
 `,
-		Args: cobra.ExactArgs(1),
 		RunE: func(c *cobra.Command, args []string) error {
 			ctx := c.Context()
 			if ctx == nil {
@@ -429,145 +459,167 @@ Examples:
 			if err != nil {
 				return err
 			}
-			switch level {
-			case "", "size", "sha256":
-			default:
-				return fmt.Errorf("--level is size or sha256, not %q", level)
-			}
-			li, err := openLocal(ctx, f, args[0])
+			li, err := target.open(ctx, f, args)
 			if err != nil {
 				return err
 			}
-			q := url.Values{}
-			if force {
-				q.Set("force", "true")
-			}
-			if level != "" {
-				q.Set("level", level)
-			}
-			path := withQuery(epLocalRetry, q)
-			var accepted struct {
-				Status        string `json:"status"`
-				PreviousPhase string `json:"previous_phase"`
-			}
-			if err := li.client.doJSON(ctx, "POST", path, nil, &accepted); err != nil {
-				return retryErr(err)
+			var cfg map[string]any
+			if err := li.client.doJSON(ctx, "GET", epLocalConfig, nil, &cfg); err != nil {
+				return err
 			}
 			if format == FormatJSON {
-				return printJSON(os.Stdout, accepted)
+				return printJSON(os.Stdout, cfg)
 			}
-			what := "re-entering the loop"
-			if force {
-				what = "re-downloading every file"
-			}
-			_, err = fmt.Printf("%s: %s, from phase %s.\n`olares-cli router local progress %s --watch` follows it.\n",
-				li.AppName, what, nonEmpty(accepted.PreviousPhase), li.AppName)
-			return err
+			return renderFlatMap(os.Stdout, cfg)
 		},
 	}
-	cmd.Flags().BoolVar(&force, "force", false, "re-download every file, ignoring the manifest")
-	cmd.Flags().StringVar(&level, "level", "", "verification for this attempt: size or sha256")
+	target.bind(cmd)
 	addOutputFlag(cmd, &output)
 	return cmd
 }
 
-func retryErr(err error) error {
-	var re *RouterError
-	if err == nil || !errors.As(err, &re) {
-		return err
-	}
-	switch {
-	case re.Status == 429:
-		return fmt.Errorf("%w\nThe application limits how often this can be asked. "+
-			"Waiting a second or two is enough; the retry it is already doing is not affected", err)
-	case re.Status == 503:
-		return fmt.Errorf("%w\nThis application has no lifecycle to retry — it is not managing a download "+
-			"at all", err)
-	}
-	return err
+// renderFlatMap prints an unknown-shaped document as sorted key/value lines,
+// descending into nested objects with dotted keys. The configuration's fields
+// are the Model Console's own and will grow; enumerating them here would mean
+// silently dropping whatever it learns to report next.
+func renderFlatMap(w io.Writer, doc map[string]any) error {
+	t := newTable(w)
+	writeFlat(t, "", doc)
+	return t.flush()
 }
 
-func newLocalRestartCommand(f *cmdutil.Factory) *cobra.Command {
-	var (
-		output string
-		yes    bool
-	)
-	cmd := &cobra.Command{
-		Use:   "restart <app>",
-		Short: "relaunch the engine with the current model card",
-		Long: `Relaunch a model application's inference engine.
-
-The weights stay where they are; only the engine process is replaced, using the
-engine arguments in the card as it stands now. That is what this is for: after
-"router local spec set", the changes an engine reads at launch — its mode, and
-capabilities its own behaviour depends on — need one.
-
-In-flight requests to that engine end. The model is then unavailable for as long
-as it takes to load the weights again, which for a large model on a cold cache
-is minutes rather than seconds, and Router reports it as unreachable meanwhile.
-
-Examples:
-  olares-cli router local restart llamacppqwen3627bggufv3
-  olares-cli router local restart llamacppqwen3627bggufv3 --yes
-`,
-		Args: cobra.ExactArgs(1),
-		RunE: func(c *cobra.Command, args []string) error {
-			ctx := c.Context()
-			if ctx == nil {
-				ctx = context.Background()
-			}
-			format, err := parseFormat(output)
-			if err != nil {
-				return err
-			}
-			li, err := openLocal(ctx, f, args[0])
-			if err != nil {
-				return err
-			}
-			if err := cliutil.ConfirmDestructive(os.Stderr, os.Stdin, fmt.Sprintf(
-				"Relaunch the engine in %s? In-flight requests end, and the model is unavailable "+
-					"until the weights load again.", li.AppName), yes); err != nil {
-				return err
-			}
-			var res struct {
-				OK     bool   `json:"ok"`
-				RunDir string `json:"run_dir"`
-			}
-			if err := li.client.doJSON(ctx, "POST", epLocalEngineRestart, nil, &res); err != nil {
-				return restartErr(ctx, li, err)
-			}
-			if format == FormatJSON {
-				return printJSON(os.Stdout, res)
-			}
-			_, err = fmt.Printf("%s: the engine was told to relaunch. It is unavailable until the weights "+
-				"are loaded again — `olares-cli router local status %s` says when.\n", li.AppName, li.AppName)
-			return err
-		},
+func writeFlat(t *table, prefix string, doc map[string]any) {
+	keys := make([]string, 0, len(doc))
+	for k := range doc {
+		keys = append(keys, k)
 	}
-	cmd.Flags().BoolVarP(&yes, "yes", "y", false, "do not ask for confirmation")
-	addOutputFlag(cmd, &output)
-	return cmd
-}
-
-func restartErr(ctx context.Context, li *llmInit, err error) error {
-	var re *RouterError
-	if err == nil || !errors.As(err, &re) {
-		return err
-	}
-	if re.Status == 404 {
-		if served := consoleServes(ctx, li, "POST", epLocalEngineRestart); served != nil && !*served {
-			return fmt.Errorf("this application's Model Console (%s) cannot relaunch its engine on "+
-				"request — the route arrived in a later version. Restarting the application does the "+
-				"same thing from outside: `olares-cli market restart %s`",
-				consoleVersion(li), li.AppName)
+	sort.Strings(keys)
+	for _, k := range keys {
+		name := k
+		if prefix != "" {
+			name = prefix + "." + k
+		}
+		switch v := doc[k].(type) {
+		case map[string]any:
+			if len(v) == 0 {
+				continue
+			}
+			writeFlat(t, name, v)
+		case nil:
+			continue
+		default:
+			t.row(name, fmt.Sprintf("%v", v))
 		}
 	}
-	switch re.Code {
-	case "no_engine":
-		return fmt.Errorf("%w\nThis application runs no engine of its own, so there is nothing to relaunch", err)
-	case "run_dir_unset":
-		return fmt.Errorf("%w\nThe Model Console cannot reach the supervisor that would do the relaunch. "+
-			"Restarting the application itself is the way round it: `olares-cli market restart <app>`", err)
+}
+
+type localEndpoint struct {
+	Method      string   `json:"method"`
+	Path        string   `json:"path"`
+	Category    string   `json:"category"`
+	Group       string   `json:"group,omitempty"`
+	Description string   `json:"description"`
+	Available   bool     `json:"available"`
+	Reasons     []string `json:"reasons,omitempty"`
+}
+
+func newModelEndpointsCommand(f *cmdutil.Factory, how localAddressing) *cobra.Command {
+	var (
+		output      string
+		unavailable bool
+	)
+	target := newModelTarget(how)
+	cmd := &cobra.Command{
+		Use:   "endpoints " + target.arg(),
+		Short: "which routes this deployment actually serves",
+		Long: `List the routes a model application serves, and which are switched off.
+
+A Model Console mounts different routes depending on what it is running: an
+embedding deployment has no chat completions, a translation one moves its routes
+to the host root, and the engine-native passthrough exists only for the engines
+that have one. So "the endpoint does not exist" and "the endpoint is not mounted
+in this deployment" are different answers, and this is where they are told apart
+— each unavailable route carries the reason it is absent.
+
+Examples:
+  olares-cli router model diag endpoints qwen3-4b
+  olares-cli router model diag endpoints --app llamacppqwen3627bggufv3 --unavailable
+`,
+		RunE: func(c *cobra.Command, args []string) error {
+			ctx := c.Context()
+			if ctx == nil {
+				ctx = context.Background()
+			}
+			format, err := parseFormat(output)
+			if err != nil {
+				return err
+			}
+			li, err := target.open(ctx, f, args)
+			if err != nil {
+				return err
+			}
+			var env struct {
+				EngineKind string          `json:"engine_kind"`
+				Endpoints  []localEndpoint `json:"endpoints"`
+			}
+			if err := li.client.doJSON(ctx, "GET", epLocalEndpoints, nil, &env); err != nil {
+				return err
+			}
+			rows := env.Endpoints
+			if unavailable {
+				kept := rows[:0:0]
+				for _, r := range rows {
+					if !r.Available {
+						kept = append(kept, r)
+					}
+				}
+				rows = kept
+			}
+			if format == FormatJSON {
+				return printJSON(os.Stdout, map[string]any{
+					"engine_kind": env.EngineKind,
+					"endpoints":   rows,
+				})
+			}
+			return renderLocalEndpoints(os.Stdout, env.EngineKind, rows, unavailable)
+		},
 	}
-	return err
+	target.bind(cmd)
+	cmd.Flags().BoolVar(&unavailable, "unavailable", false, "only the routes this deployment does not serve")
+	addOutputFlag(cmd, &output)
+	return cmd
+}
+
+func renderLocalEndpoints(w io.Writer, engineKind string, rows []localEndpoint, onlyUnavailable bool) error {
+	if len(rows) == 0 {
+		if onlyUnavailable {
+			_, err := fmt.Fprintln(w, "every route this Model Console knows about is mounted.")
+			return err
+		}
+		_, err := fmt.Fprintln(w, "this Model Console reported no routes at all, which should not happen.")
+		return err
+	}
+	if engineKind != "" {
+		if _, err := fmt.Fprintf(w, "engine: %s\n\n", engineKind); err != nil {
+			return err
+		}
+	}
+	sort.SliceStable(rows, func(i, j int) bool {
+		if rows[i].Category != rows[j].Category {
+			return rows[i].Category < rows[j].Category
+		}
+		return rows[i].Path < rows[j].Path
+	})
+	t := newTable(w, "METHOD", "PATH", "CATEGORY", "SERVED", "WHY NOT")
+	for _, r := range rows {
+		why := ""
+		if !r.Available {
+			why = clip(strings.Join(r.Reasons, "; "), 60)
+			if why == "" {
+				why = "no reason given"
+			}
+		}
+		t.row(r.Method, r.Path, r.Category, boolStr(r.Available), why)
+	}
+	return t.flush()
 }
