+++ cli/cmd/ctl/skills_commands_test.go
@@ -135,6 +135,15 @@ func TestSkillCommandPathsExist(t *testing.T) {
 		"router model add",
 		"router model update",
 		"router model remove",
+		"router model status",
+		"router model progress",
+		"router model retry",
+		"router model restart",
+		"router model diag",
+		"router model diag gpu",
+		"router model diag perf",
+		"router model diag config",
+		"router model diag endpoints",
 		"router route",
 		"router route list",
 		"router route get",
@@ -148,7 +157,8 @@ func TestSkillCommandPathsExist(t *testing.T) {
 		"router model spec",
 		"router model spec show",
 		"router model spec edit",
-		"router model restart",
+		"router model spec file",
+		"router model spec set",
 		"router call",
 		"router call chat",
 		"router call embed",
@@ -183,19 +193,6 @@ func TestSkillCommandPathsExist(t *testing.T) {
 		"router audit",
 		"router audit list",
 		"router audit get",
-		"router local",
-		"router local status",
-		"router local progress",
-		"router local spec",
-		"router local spec show",
-		"router local spec file",
-		"router local spec set",
-		"router local config",
-		"router local endpoints",
-		"router local gpu",
-		"router local perf",
-		"router local retry",
-		"router local restart",
 	}
 
 	root := NewDefaultCommand()
